Extracted common code for processing multipart file uploads.
