+++ api/V1/File.js
@@ -1,13 +1,10 @@
 const models            = require('../../models');
-const log               = require('../../logging');
+const util              = require('./util');
 const responseUtil      = require('./responseUtil');
-const multiparty        = require('multiparty');
 const config            = require('../../config');
-const uuidv4            = require('uuid/v4');
 const jsonwebtoken      = require('jsonwebtoken');
-const modelsUtil        = require('../../models/util/util');
 const middleware        = require('../middleware');
-const { query, check }  = require('express-validator/check');
+const { query }  = require('express-validator/check');
 
 module.exports = (app, baseUrl) => {
   var apiUrl = baseUrl + '/files';
@@ -36,92 +33,16 @@ module.exports = (app, baseUrl) => {
     [
       middleware.authenticateUser,
     ],
-    middleware.requestWrapper(async (request, response) => {
-
-      var dbRecord;
-      var key = uuidv4();
-      var validData = false;
-      var uploadStarted = false;
-
-      var form = new multiparty.Form();
-      // Make new userfile from the data field and Device.
-      // TODO Stop stream if 'data' is invalid.
-      form.on('field', (name, value) => {
-        if (name != 'data') {
-          return; // Only parse data field.
-        }
-        try {
-          var data = JSON.parse(value);
-          dbRecord = models.File.build(data, {
-            fields: models.File.apiSettableFields,
-          });
-          dbRecord.set('fileKey', key);
-          dbRecord.set('UserId', request.user.id);
-          validData = true;
-        } catch (e) {
-          log.debug(e);
-          // TODO Stop stream here.
-        }
-      });
-
-      // Stream file to S3.
-      var uploadPromise;
-      form.on('part', (part) => {
-        if (part.name != 'file') {
-          return part.resume();
-        }
-        uploadStarted = true;
-        log.debug('Streaming file to bucket.');
-        uploadPromise = new Promise(function(resolve, reject) {
-          var s3 = modelsUtil.openS3();
-          s3.upload({
-            Bucket: config.s3.bucket,
-            Key: key,
-            Body: part,
-          }, (err) => {
-            if (err) {
-              return reject(err);
-            }
-            log.info("Finished streaming file to object store key:", key);
-            resolve();
-          });
+    middleware.requestWrapper(
+      util.multipartDownload('file', (request, data, key) => {
+        var dbRecord = models.File.build(data, {
+          fields: models.File.apiSettableFields,
         });
-      });
-
-      // Close response.
-      form.on('close', async () => {
-        log.info("Finished POST request.");
-        if (!validData || !uploadStarted) {
-          return responseUtil.invalidDatapointUpload(response);
-        }
-
-        // Check that the file uploaded to file store.
-        try {
-          await uploadPromise;
-        } catch (e) {
-          return responseUtil.send(response, {
-            statusCode: 500,
-            success: false,
-            messages: ["Failed to upload file to bucket"],
-          });
-        }
-        // Validate and upload recording
-        await dbRecord.validate();
-        await dbRecord.save();
-        return responseUtil.validRecordingUpload(response, dbRecord.id);
-      });
-
-      form.on('error', (e) => {
-        log.error(e);
-        return responseUtil.send(response, {
-          statusCode: 400,
-          success: false,
-          messages: ["failed to get file content"],
-        });
-      });
-
-      form.parse(request);
-    })
+        dbRecord.set('UserId', request.user.id);
+        dbRecord.set('fileKey', key);
+        return dbRecord;
+      })
+    )
   );
 
   /**
@@ -129,7 +50,7 @@ module.exports = (app, baseUrl) => {
    * @apiName GetFiles
    * @apiGroup Files
    *
-   * @apiUse V1UserAuthorizationHeader
+   * @apiHeader {String} Authorization Signed JSON web token for a user or device.
    *
    * @apiParam {JSON} where [Sequelize where conditions](http://docs.sequelizejs.com/manual/tutorial/querying.html#where) for query.
    * @apiParam {Number} offset Query result offset (for paging).
@@ -148,7 +69,7 @@ module.exports = (app, baseUrl) => {
   app.get(
     apiUrl,
     [
-      middleware.authenticateUser,
+      middleware.authenticateIsFromSite,
       middleware.parseJSON('where'),
       query('offset').isInt().optional(),
       query('limit').isInt().optional(),
@@ -232,4 +153,4 @@ module.exports = (app, baseUrl) => {
       });
     })
   );
- };
+};
